+++ Battler.rb
@@ -4807,7 +4807,7 @@ class PokeBattle_Battler
       user.pbCancelMoves
       return false
     end
-    if !basemove.zmove # Z-Moves handle protection stuff elsewhere
+    if !basemove.zmove && !basemove.giga # Z-Moves handle protection stuff elsewhere
       if target.pbOwnSide.effects[:MatBlock] && [:PHANTOMFORCE, :SHADOWFORCE, :HYPERSPACEHOLE, :HYPERSPACEFURY].include?(basemove.move)
         @battle.pbDisplay(_INTL("The Mat Block was broken!"))
       end
@@ -4883,7 +4883,7 @@ class PokeBattle_Battler
             user.pbBurn(target)
             @battle.pbDisplay(_INTL("{1}'s Burning Bulwark burned {2}!",target.pbThis,user.pbThis(true)))
           end
-        end
+        end 
         return false
       end
     end
@@ -5049,6 +5049,22 @@ class PokeBattle_Battler
         self.pbIncreaseStat(PBStats::ATTACK,1,abilitymessage:false)
         self.pbIncreaseStat(PBStats::SPATK,1,abilitymessage:false)
         @battle.pbDisplay(_INTL("{1} switched to Attack Stance!",pbThis))
+      end
+    end
+    if (self.crested == :CACTURNE)
+      cacturne_first = true
+      for i in @battle.battlers
+        next if i.isFainted? || !pbIsOpposing?(i.index)
+        if (i.hasMovedThisRound?)
+          cacturne_first = false;
+        end
+      end
+
+      if cacturne_first && (!self.pbTooHigh?(PBStats::ATTACK) || !self.pbTooHigh?(PBStats::SPATK)) && basemove.category != :status
+        @battle.pbCommonAnimation("StatUp",self,nil)
+        self.pbIncreaseStatBasic(PBStats::ATTACK,1)
+        self.pbIncreaseStatBasic(PBStats::SPATK,1)
+        @battle.pbDisplay(_INTL("{1}'s crest raised its offenses!",self.pbThis))
       end
     end
      # Stance Change moved from here to end of method to match Gen VII mechanics.